// dev-admin/src/app/theme/vendors/vendors-add/vendors-add.interface.ts
export interface VendorInetrface {
    userId: string;
    vendorName: string;
    boutiqueName: string;
    storeName: string;
    priceId: string;
    email: string;
    phoneNumber: string;
    address1: string;
    address2: string;
    city: string;
    state: string;
    country: string;
    zip: string;
    shippingRate: string;
    returnShippingRate: string;
    status?:string;
}